fix: more correct DB and user name handling (#580)
PostgreSQL requires that the DB name is a valid identifier

https://www.postgresql.org/docs/current/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS

This was causing some issues with metrics gathering in the past, but
that should be resolved now. This makes it more in-line with what
PostgreSQL is allowing and expecting. The max length is compile-time
configurable, as PostgreSQL technically provides a way to increase the
byte limit, by default we use current PostgreSQL default there.

This mean that some users *may* be allowed to use weird names, but
assuming that Victoria Metrics is properly parsing exported metrics
using OpenMetrics specs, then it should work there.

Close #579