fix: lower log level when removing stopped listener (#577)
This message is not an error, and it needlessly scares operators. This
error seems like it is bug in Ranch, as according to the docs there
should be no error and instead we should get `{:error, :not_found}`
response, but for now let's deal with it by lowering the error message
to make it simpler.