fix: do not handle errors during connection
This is potential source of a problem, as if there are is error in 2nd
or 3rd clause of `with` then we are leaking connections which in the end
may result in the connections exhaustion. Instead we should fail loudly
that there is something wrong with the connections. It should also help
with debugging potential problems.