feat: improve metrics gathering
Instead of fetching metrics results as a binary and then doing some
magic to add global tags and to merge them, merge them in their basic
form as produced by the `Peep` and then generate report from that on
target node. This should resolve a lot of problems related to the output
format issues.