feat: JSON output format for log messages
Output logs in line JSON format (JSON object per line) in format that is
as similar to the format that is expected by the Logflare as possible.

The idea there is to log to the Vector which will then take care of
logging to file(s) and to dispatch the logs to Logflare.