fix(logger): format mfas in accordance to logflare, put level in meta (#667)
## What kind of change does this PR introduce?

Changes file log formatting to conform to Logflare's format.

## What is the current behavior?

Because MFAs have a mix of strings and integers, ingestion fails. Also,
the log level is on top level, but logflare expects it to be inside
metadata, so it wouldn't get picked up.

## What is the new behavior?

Added a pass to perform arbitrary transformations.

## Additional context

Tested in staging