fix: don't stop client connection when parameter status is updated (#666)
Instead of stopping the client connection, just continue with the
greeting process, with the new parameter status.

Also adds a test script that performs a pseudo postgres update. In the
original code, the query must be retried on Supavisor because the first
connection attempt fails. After the patch, no retries are needed.

Discovered in #663.