fix: close proxy websocket upon TCP connection closing (#684)
This fixes a bug where clients using `neondatabase/serverless` would
have their code hang for 60s upon calling `client.end()`. Essentially
the client closes the (proxied) TCP socket, and we need to ensure that
the websocket is closed as well. Otherwise the client keeps blocked
waiting for the websocket to terminate.

Added an integration test with such a case (times out without the
patch). Unfortunately, the test requires bumping node, so I did that as
well, will check to ensure it doesn't impact anything else.