fix(client-handler): handle :cancel_query when state is :idle (#692)
Fixes
```
[error] ClientHandler: Undefined msg: [{"type", :info}, {"content", :cancel_query}]
```
I see it pretty often in the logs. It seems to me that it should be just
a no-op (query cancel coming late, when the query was already
completed). The current code already ignores it, so the patch
effectively just removes the error log (and adds a dedicated code path
for it)