perf: deterministic prepared statement names
Using phash2 in the query string to determine the name. This allows us
to avoid preparing the same query twice, when they get routed from
different client connections to the same db_handler.

The approach to inject the parse response is subject to change, as
it doesn't work well if the client doesn't use pipelining.

Tests pending as well.