feat: refresh PromEx application metrics on soft deploy
The PromEx application metrics are "manual", so they are only
emitted on application start or upon calling
`PromEx.ManualMetricsManager.refresh_metrics/1`. Since we don't do
a restart on soft deploys, the relevant metrics (application version)
stays outdated in Grafana until a restart happens. This commit
adds a refresh in the update script.