feat: release hooks
The functions from this module should be called at specific points
in the release process, and may interfere with the results. Can
perform actions for which the appups aren't adequate. Currently,
most of the functions don't do much: they are there so that, in future
versions, they can be modified without changes in external tooling.