feat: add authentication_cleartext_password support (#707)
## What kind of change does this PR introduce?

Feature

## What is the current behavior?

db_handler understands the authentication_cleartext_password message but
doesn't have a handler. This means servers using PAM for authentication
do not work with db_handler as the password is never sent in response.

## What is the new behavior?

Adds a handler for authentication_cleartext_password

## Additional context

This is needed for other projects that will use PAM to extend the auth
mechanisms offered in the hosted postgres at Supabase

---------

Co-authored-by: felipe stival <14948182+v0idpwn@users.noreply.github.com>