fix(authn): when credential changes are detected, retry the auth request  (#728)
- Fixes an issue where the first authn request after a credential change
always would fail.
- Adds a cache refresh limit of 3 per minute per pool (triggered every
time the password is wrong). Previosuly, limited to once every 15s. This
is mostly due to tests. In tests, if we don't have this, the hard limit
of once per 15s causes the test to flak.
- Adds an optimization where we use the existing SecretsChecker
connection instead of opening a new one to refresh the cache.