fix: always terminate db_handler when owner client_handler terminates
Among other things, this patch addresses a race condition when the
socket was closed while DbHandler had `active: false`. This led to the
DbHandler not receiving the TCP closed message, and the client handler
crashing due to it. The DbHandler would stay on the pool with the dead
socket, and and successive checkouts would get an unusable connection.

This patch ensures that the DbHandler is closed whenever the owner
terminates, regardless of how it terminates or pool mode. This may cause
the transaction pooler to restart connections more often, but should give
us more correct behaviour.