chore: add upgrade_previous
Upgrade previous should be used to generate the appups, etc. While upgrade_from should be used to refer to what's the base version installed

If going in a install path such as 2.1.3 --soft--> 2.1.7 --soft--> 2.1.11, in the first soft update upgrade_from/upgrade_previous will be 2.1.3. In the second one, upgrade_from will be 2.1.3, but upgrade_previous will be 2.1.7