fix: session query metrics (#801)
In v2, session mode query metrics weren't tracked properly. Due to
`DbHandler` not calling `db_status/1` it never notified `ClientHandler`
that a transaction terminated, meaning `ClientHandler` wouldn't ever
collect the query time metric.

Additionally, this adds a `:proxy` metadata to queries, indicating
whether the query was proxied or not: all queries emit `duration` twice:
one in the proxy node, another in the pool node. Can be useful
to compare proxy performance vs unproxied, and detect proxying-caused
degradations. Another option for us is to not emit the query duration on
the pool owner node.