fix: ensure error is sent when db handler terminates during checkout
This happens often in proxy and session modes in v2.8. The ClientHandler
that is trying to check-out the process gets the termination, not a
response, and just crashes, terminating the socket.

This fix ensures that DbHandler sends a reply to the checkout and we
write the error to the socket when we are terminating due to initialization
errors.

The approach consists in introducing a new state. Since the checkout
message is postponed in the auth state, when the statem is transitioned
to the `:terminating_with_error` state we have an opportunity to handle
it again.

This PR also changes a bit how proxy is handled. Instead of having the
caller/client_sock from the beginning, the behaviour is now closer
to the session mode, where we get the client socket after the checkout.