fix(db-handler): use {active, N} instead of {active, true} after receiving ReadyForQuery (#816)
The usage of `active: true` when receiving a ReadyForQuery message
hindered the expected backpressure on DbHandler, allowing it to
accumulate a high number of TCP messages in queue, taking a lot
of memory.

Since the large binaries are off process heap, it didn't flag large
heap, so
the process wouldn't be terminated due to too much mem.